+++ Cargo.toml
@@ -4,6 +4,7 @@ members = [
     "packages/dioxus",
     "packages/dioxus-lib",
     "packages/core",
+    "packages/core-types",
     "packages/cli",
     "packages/cli-config",
     "packages/core-macro",
@@ -68,6 +69,7 @@ version = "0.6.0-alpha.2"
 dioxus = { path = "packages/dioxus", version = "0.6.0-alpha.0" }
 dioxus-lib = { path = "packages/dioxus-lib", version = "0.6.0-alpha.0" }
 dioxus-core = { path = "packages/core", version = "0.6.0-alpha.0" }
+dioxus-core-types = { path = "packages/core-types", version = "0.6.0-alpha.0" }
 dioxus-core-macro = { path = "packages/core-macro", version = "0.6.0-alpha.0" }
 dioxus-config-macro = { path = "packages/config-macro", version = "0.6.0-alpha.0" }
 dioxus-router = { path = "packages/router", version = "0.6.0-alpha.0" }
@@ -76,6 +78,7 @@ dioxus-html = { path = "packages/html", version = "0.6.0-alpha.0", default-featu
 dioxus-html-internal-macro = { path = "packages/html-internal-macro", version = "0.6.0-alpha.0" }
 dioxus-hooks = { path = "packages/hooks", version = "0.6.0-alpha.0" }
 dioxus-web = { path = "packages/web", version = "0.6.0-alpha.0", default-features = false }
+dioxus-isrg = { path = "packages/isrg", version = "0.6.0-alpha.0" }
 dioxus-ssr = { path = "packages/ssr", version = "0.6.0-alpha.0", default-features = false }
 dioxus-desktop = { path = "packages/desktop", version = "0.6.0-alpha.0", default-features = false }
 dioxus-mobile = { path = "packages/mobile", version = "0.6.0-alpha.0" }
@@ -102,6 +105,7 @@ warnings = { version = "0.2.0" }
 # a fork of pretty please for tests
 prettier-please = { version = "0.3.0", features = ["verbatim"]}
 
+askama_escape = "0.10.3"
 tracing = "0.1.37"
 tracing-futures = "0.2.5"
 toml = "0.8"
