+++ packages/interpreter/Cargo.toml
@@ -24,6 +24,7 @@ serde = { version = "1.0", features = ["derive"], optional = true }
 rustc-hash = { workspace = true, optional = true }
 
 dioxus-core = { workspace = true, optional = true }
+dioxus-core-types = { workspace = true, optional = true }
 dioxus-html = { workspace = true, optional = true }
 
 [build-dependencies]
@@ -41,5 +42,5 @@ webonly = [
     "dep:web-sys",
     "sledgehammer_bindgen/web",
 ]
-binary-protocol = ["sledgehammer", "dep:dioxus-core", "dep:dioxus-html"]
+binary-protocol = ["sledgehammer", "dep:dioxus-core", "dep:dioxus-core-types", "dep:dioxus-html"]
 minimal_bindings = []
