+++ packages/ssr/Cargo.toml
@@ -10,42 +10,12 @@ keywords = ["dom", "ui", "gui", "react", "ssr"]
 
 [dependencies]
 dioxus-core = { workspace = true, features = ["serialize"] }
-dioxus-html = { workspace = true, features = ["document"]}
-dioxus-cli-config = { workspace = true, features = ["read-config"], optional = true }
-dioxus-interpreter-js = { workspace = true }
-generational-box = { workspace = true }
-askama_escape = "0.10.3"
-thiserror = "1.0.23"
-rustc-hash = "1.1.0"
-lru  = { workspace = true }
-tracing = { workspace = true }
-http = { workspace = true }
-async-trait = { workspace = true }
-serde_json = { workspace = true }
-chrono = { version = "0.4.34", optional = true }
-
-[target.'cfg(target_arch = "wasm32")'.dependencies]
-tokio = { version = "1.28", features = ["io-util"], optional = true }
-
-[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
-tokio = { version = "1.28", features = ["fs", "io-util"], optional = true }
+dioxus-core-types = { workspace = true }
+askama_escape = { workspace = true }
+rustc-hash = { workspace = true }
 
 [dev-dependencies]
 dioxus = { workspace = true }
-dioxus-signals = { workspace = true }
-tokio = { version = "1", features = ["full"] }
-
-tracing = { workspace = true }
-fern = { version = "0.6.0", features = ["colored"] }
-anyhow = "1.0"
-argh = "0.1.4"
-serde = "1.0.120"
-serde_json = "1.0.61"
-fs_extra = "1.2.0"
-
-[features]
-default = []
-incremental = ["dep:tokio", "dep:chrono", "dep:dioxus-cli-config"]
 
 [package.metadata.docs.rs]
 cargo-args = ["-Zunstable-options", "-Zrustdoc-scrape-examples"]
