+++ packages/ssr/src/renderer.rs
@@ -243,7 +243,7 @@ impl Renderer {
                     // then write any listeners
                     for name in accumulated_listeners.drain(..) {
                         write!(buf, ",{}:", &name[2..])?;
-                        write!(buf, "{}", dioxus_html::event_bubbles(&name[2..]) as u8)?;
+                        write!(buf, "{}", dioxus_core_types::event_bubbles(&name[2..]) as u8)?;
                     }
                 }
