+++ packages/static-generation/Cargo.toml
@@ -14,7 +14,8 @@ resolver = "2"
 dioxus-fullstack = { workspace = true }
 dioxus-lib.workspace = true
 dioxus-router = { workspace = true, features = ["fullstack"]}
-dioxus-ssr = { workspace = true, features = ["incremental"], optional = true }
+dioxus-ssr = { workspace = true, optional = true }
+dioxus-isrg = { workspace = true, optional = true }
 axum = { workspace = true, features = ["ws", "macros"], optional = true }
 tower-http = { workspace = true, features = ["fs"], optional = true }
 dioxus-hot-reload = { workspace = true, features = ["serve"], optional = true }
@@ -33,11 +34,6 @@ criterion = { workspace = true }
 default = []
 server = ["dioxus-fullstack/server", "dioxus-router/ssr", "dep:dioxus-ssr", "dep:tokio", "dep:http", "dep:axum", "dep:tower-http", "dep:dioxus-hot-reload", "dep:dioxus-cli-config", "dep:tower"]
 web = ["dioxus-fullstack/web", "dioxus-router/web", "dep:dioxus-web"]
-
-# [[bench]]
-# name = "incremental"
-# required-features = ["server"]
-# harness = false
 
 [package.metadata.docs.rs]
 cargo-args = ["-Zunstable-options", "-Zrustdoc-scrape-examples"]
