+++ packages/web/Cargo.toml
@@ -11,6 +11,7 @@ keywords = ["dom", "ui", "gui", "react", "wasm"]
 
 [dependencies]
 dioxus-core = { workspace = true }
+dioxus-core-types = { workspace = true }
 dioxus-html = { workspace = true, features = ["wasm-bind"] }
 dioxus-hot-reload = { workspace = true, features = ["client"] }
 dioxus-signals = { workspace = true }
